// src/app/models/parametro.ts
export class Parametro {
    id: number;
    nombre: string;
    variable1anexo: string;
    variable2anexo: string;
    variable3anexo: string;
    variable4anexo: string;
    variable5anexo: string;
    metodo: string;
    tecnica_analitica: string;
    unidad: string;
    valor_unit: number;
    logicalerasure: Boolean;
    descripcion: string;
    codigo: string;
    variables: any;
    formula: String;
    validaciones: any;
    plataformaTecnica: Boolean;
    cuentaActiva: Boolean;
    plataformaAdmin: Boolean;
    admin: Boolean;
    index: number;
    rol: string;
}